Randomize order of locations within a bucket
OK, so this is a little bit tricky. The order of locations within a bucket
should be randomized. However, it should stay consistent throughout the
entire 5pm hour, because the javascript might be involved, so it should
get the same result. So, use the Moment.valueOf of the 5:00:00 pm instant
of a bucket (regardless of whether its 5:00pm or 5:37pm or 5:59:59pm) to
seed a predictable random generator (Merseinne Twister).