Create tweet details page
Create a page that, for any Twitter ID, shows that tweet and the places in the
world where it was 5pm.

Because this page will use the URL scheme /t/:tweet_id, I've moved the page
that shows the standard homepage for any given time to /d/:iso_date. Not that
it matters, because that URL is not publicized, and may never end up being
used.